requirements.txt
PyYAML==6.0
py-cord==2.4.1
